+++ src/EnergyModelsRecHorizon.jl
@@ -13,6 +13,7 @@ using CSV
 const EMB = EnergyModelsBase
 const TS = TimeStruct
 
+include(joinpath("structures", "horizons.jl"))
 include("datastructures.jl")
 include("model.jl")
 include("utils.jl")
@@ -28,6 +29,10 @@ export InitData, InitStorageData
 
 export has_init, is_init_data
 
-# export previous_level
+# Types utilized in the instantiation of a model
+export DurationHorizons, PeriodHorizons
+
+# Functions for accessing the fields of `SingleHorizon` instances
+export durations, indices_optimization, indices_implementation
 
 end # module EnergyModelsRecHorizon