+++ test/test_examples.jl
@@ -1,18 +1,14 @@
-"""
-Running examples folder
-"""
+ENV["EMX_TEST"] = true # Set flag for example scripts to check if they are run as part of the tests
 
 @testset "Run examples" begin
-    exdir = joinpath(@__DIR__, "../examples")
-    files = first(walkdir(exdir))[3]
-    for file in files
-        if splitext(file)[2] == ".jl"
-            @testset "Example $file" begin
-                @info "Run example $file"
+    exdir = joinpath(@__DIR__, "..", "examples")
+    files = filter(endswith(".jl"), readdir(exdir))
+    for file ∈ files
+        @testset "Example $file" begin
+            redirect_stdio(stdout=devnull, stderr=devnull) do
                 include(joinpath(exdir, file))
-
-                @test termination_status(m) == MOI.OPTIMAL
             end
+            @test termination_status(m) == MOI.OPTIMAL
         end
     end
     Pkg.activate(@__DIR__)