+++ NEWS.md
@@ -1,5 +1,17 @@
 # Release notes
 
+
+## Unversioned
+
+### Time iterators for the horizon
+
+* Included new types for passing arguments for creating the iterators for the receding horizon framework:
+  * `PeriodHorizons` defines the horizons by the number of operational periods,
+  * `DurationHorizons` defines the horizons by the duration of the operational periods.
+* Both types have an iterator in which the element is provided by a `SingleHorizon`:
+  * it includes all information required for setting up a single horizon,
+  * it allows for creating the individual models
+
 ## Version 0.2.0 (2024-08-07)
 
 ### First working version
@@ -8,7 +20,7 @@
 slice of time from the original problem. This approach will be replaced in the
 future.
 * Introduces `InitData` as abstract type for initialization data.
-* Introduces `RecHorEnergyModel` for dispatching functions that are relevant for 
+* Introduces `RecHorEnergyModel` for dispatching functions that are relevant for
 initialization.
 
 ## Version 0.1.0 (2024-04-11)